Javascript structure refactoring: Second version relates to xibosignageltd/xibo-private#842
 - xibo-cms.js datatables moved to a new file
 - authed.twig and base.twig translations, variables and configs moved to dedicated files
  - translations and vars for editors moved to dedicated files
 - codemirror replaces heavier monaco code editor
 - old region forms removed
 - developer template page migrated
 - region and widget edit request changed to not display form
 - region and widget delete form route removed
 - file upload js is bundled
