+++ lib/Controller/Widget.php
@@ -310,14 +310,14 @@ class Widget extends Base
     }
 
     /**
-     * Edit Widget Form
+     * Get Widget
      * @param Request $request
      * @param Response $response
      * @param $id
      * @return \Psr\Http\Message\ResponseInterface|Response
      * @throws \Xibo\Support\Exception\GeneralException
      */
-    public function editWidgetForm(Request $request, Response $response, $id)
+    public function getWidget(Request $request, Response $response, $id)
     {
         // Load the widget
         $widget = $this->widgetFactory->loadByWidgetId($id);
@@ -595,44 +595,6 @@ class Widget extends Base
             'message' => sprintf(__('Edited %s'), $module->name),
             'id' => $widget->widgetId,
             'data' => $widget
-        ]);
-
-        return $this->render($request, $response);
-    }
-
-    /**
-     * Delete Widget Form
-     * @param Request $request
-     * @param Response $response
-     * @param $id
-     * @return \Psr\Http\Message\ResponseInterface|Response
-     * @throws AccessDeniedException
-     * @throws GeneralException
-     * @throws NotFoundException
-     * @throws \Xibo\Support\Exception\ControllerNotImplemented
-     */
-    public function deleteWidgetForm(Request $request, Response $response, $id)
-    {
-        $widget = $this->widgetFactory->loadByWidgetId($id);
-
-        if (!$this->getUser()->checkDeleteable($widget)) {
-            throw new AccessDeniedException(__('This Widget is not shared with you with delete permission'));
-        }
-
-        $error = false;
-        $module = null;
-        try {
-            $module = $this->moduleFactory->getByType($widget->type);
-        } catch (NotFoundException $notFoundException) {
-            $error = true;
-        }
-
-        // Pass to view
-        $this->getState()->template = 'module-form-delete';
-        $this->getState()->setData([
-            'widgetId' => $id,
-            'module' => $module,
-            'error' => $error,
         ]);
 
         return $this->render($request, $response);
