+++ package.json
@@ -47,7 +47,6 @@
     "imports-loader": "^5.0.0",
     "less": "^4.2.0",
     "less-loader": "^12.2.0",
-    "monaco-editor-webpack-plugin": "^7.1.0",
     "postcss-loader": "^8.1.1",
     "sass": "^1.77.8",
     "sass-loader": "^16.0.0",
@@ -55,13 +54,18 @@
     "style-loader": "^4.0.0",
     "url-loader": "^4.1.1",
     "webpack": "^5.93.0",
-    "webpack-bundle-analyzer": "^4.10.2",
     "webpack-cli": "^5.1.4"
   },
   "dependencies": {
+    "@codemirror/lang-css": "^6.3.0",
+    "@codemirror/lang-html": "^6.4.9",
+    "@codemirror/lang-javascript": "^6.2.2",
+    "@codemirror/state": "^6.4.1",
     "@fortawesome/fontawesome-free": "^6.6.0",
     "@mapbox/leaflet-pip": "^1.1.0",
     "@popperjs/core": "^2.11.8",
+    "@ssddanbrown/codemirror-lang-twig": "^1.0.0",
+    "@xiechao/codemirror-lang-handlebars": "^1.0.4",
     "ajax-bootstrap-select": "^1.4.5",
     "blueimp-file-upload": "^10.32.0",
     "blueimp-load-image": "~5.16.0",
@@ -75,6 +79,7 @@
     "bootstrap-tagsinput": "^0.7.1",
     "chart.js": "^2.9.4",
     "chartjs-plugin-datalabels": "1.0.0",
+    "codemirror": "^6.0.1",
     "colors.js": "~1.2.4",
     "corejs-typeahead": "^1.3.3",
     "datatables.net": "^1.11.3",
@@ -108,7 +113,6 @@
     "moment": "~2.30.1",
     "moment-jalaali": "~0.10.1",
     "moment-timezone": "^0.5.45",
-    "monaco-editor": "^0.50.0",
     "moveable": "^0.53.0",
     "pdfjs-dist": "2.4.456",
     "persian-date": "^1.1.0",
