+++ ui/bundle_code_editor.js
@@ -1,0 +1,40 @@
+/*
+ * Copyright (C) 2024 Xibo Signage Ltd
+ *
+ * Xibo - Digital Signage - https://xibosignage.com
+ *
+ * This file is part of Xibo.
+ *
+ * Xibo is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * any later version.
+ *
+ * Xibo is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with Xibo.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+// CodeMirror
+import {basicSetup} from 'codemirror';
+import {EditorView, keymap} from '@codemirror/view';
+import {indentWithTab} from '@codemirror/commands';
+import {javascript} from '@codemirror/lang-javascript';
+import {css} from '@codemirror/lang-css';
+import {html} from '@codemirror/lang-html';
+import {twig} from '@ssddanbrown/codemirror-lang-twig';
+
+window.CodeMirror = {
+  basicSetup,
+  EditorView,
+  keymap,
+  indentWithTab,
+  javascript,
+  css,
+  html,
+  twig,
+};
