+++ ui/bundle_editor_common.js
@@ -1,3 +1,4 @@
+/* eslint-disable no-unused-vars */
 /*
  * Copyright (C) 2024 Xibo Signage Ltd
  *
