+++ ui/bundle_xibo.js
@@ -28,5 +28,11 @@ require('./src/core/forms.js');
 require('./src/core/help-pane.js');
 require('./src/style/help-pane.scss');
 
+// Xibo calendar
+require('./src/core/xibo-calendar.js');
+
 // Xibo core
 require('./src/core/xibo-cms.js');
+
+// Xibo datatables and folders
+require('./src/core/xibo-datatables.js');
