+++ ui/src/style/forms.scss
@@ -48,21 +48,20 @@
 /* Code input */
 .code-input-group {
     .code-input-editor-container {
-        background-color: $xibo-color-neutral-900;
-        padding: 6px;
-        height: 250px;
-        border: 2px solid $xibo-color-neutral-700;
+        min-height: 160px;
+        max-height: 400px;
+        border: 2px solid $xibo-color-neutral-600;
         border-radius: 0.25rem;
+        overflow-y: auto;
 
         .code-input-editor {
             width: 100%;
             height: 100%;
         }
-    }
 
-    /* Fix for monaco editor helper */
-    .monaco-editor .editor-widget.suggest-widget {
-        left: 0 !important;
+        .cm-gutters {
+            display: none;
+        }
     }
 }
