+++ ui/src/templates/forms/region.hbs
@@ -4,7 +4,6 @@
         {{#if showExitTransition}}
             <li class="nav-item"><a class="nav-link" href="#transitionTab" role="tab" data-toggle="tab"><span>{{trans.transition}}</span></a></li>
         {{/if}}
-        <li class="nav-item"><a class="nav-link" href="#positionTab" role="tab" data-toggle="tab"><i class="fas fa-border-none tooltip-always-on" data-toggle="tooltip" data-title="{{trans.positioning}}"></i></a></li>
     </ul>
     <div class="form-content">
         <form id="regionFormEdit" class="XiboForm">
@@ -30,16 +29,6 @@
                         {{> inputs/dropdown value=(getOption region.regionOptions "transitionType") options=(arrMerge (arr (obj code='' transition='Off')) transitions.out) optionsTitle="transition" optionsValue="code" id="transitionType" title=trans.region.exitTransitionTitle helpText=trans.region.exitTransitionHelpText }}
                     </div>
                 {{/if}}
-
-                <div class="tab-pane" id="positionTab">
-                    {{> button id="bringToViewButton" type="btn-block btn-outline-success inline-btn d-none mb-2" name=trans.region.bringRegionToView customClass="position-input" }}
-                    {{> inputs/number value=region.zIndex id="zIndex" title=trans.region.layer helpText=trans.region.layerHelpText customClass="position-input" }}
-                    {{> inputs/number value=region.top id="top" title=trans.region.top helpText=trans.region.topHelpText customClass="position-input" }}
-                    {{> inputs/number value=region.left id="left" title=trans.region.left helpText=trans.region.leftHelpText customClass="position-input" }}
-                    {{> inputs/number value=region.width id="width" title=trans.region.width helpText=trans.region.widthHelpText customClass="position-input" }}
-                    {{> inputs/number value=region.height id="height" title=trans.region.height helpText=trans.region.heightHelpText customClass="position-input" }}
-                    {{> button id="setFullScreenButton" type="btn-block btn-outline-primary inline-btn" name=trans.region.makeRegionFullScreen customClass="position-input" }}
-                </div>
             </div>
         </form>
     </div>
