+++ views/authed.twig
@@ -136,92 +136,6 @@
 {% endblock %}
 
 {% block javaScriptTemplates %}
-    <script type="text/javascript" nonce="{{ cspNonce }}">
-        var userPreferencesUrl = "{{ url_for("user.pref") }}";
-
-        // TODO: Change this to a call to all route permissions ( something like currentUser.getRoutePermissions() )
-        var userRoutePermissions = {
-            tags: "{{ currentUser.featureEnabled('tag.tagging') }}"
-        };
-
-        var tagGetByNameUrl = "{{ url_for("tag.getByName") }}";
-        var tagEditMultipleUrl = "{{ url_for("tag.editMultiple", {type: '[type]'}) }}";
-        var tagSearchUrl = "{{ url_for("tag.search") }}";
-    </script>
-
-    <!-- TODOM move this JS??? -->
-    <script type="text/javascript" nonce="{{ cspNonce }}">
-        $(document).on('click', '#dropdown-menu-right-container a.dropdown-item', function(e) {
-            let eventType = $(this).attr('data-custom-handler');
-            let eventParam = $(this).attr('data-custom-handler-url');
-
-            if (eventType === 'createMiniLayoutPreview') {
-                $(this).click(createMiniLayoutPreview(eventParam));
-            }
-        })
-
-        function moveFolderMultiSelectFormOpen(dialog) {
-            // make bootstrap happy.
-            if ($('#folder-tree-form-modal').length != 0) {
-                $('#folder-tree-form-modal').remove();
-            }
-
-            var folderContainer = '<div class="card p-3 mb-3 bg-light" id="container-folder-form-tree"></div>';
-            var $hiddenInput = $('<input name="folderId" type="hidden" id="formFolderId">');
-
-            $hiddenInput.on('change', function() {
-                dialog.data().commitData = {folderId: $(this).val()};
-            });
-
-            $(dialog).find('.modal-body').append(folderContainer);
-            $(dialog).find('.modal-body').append($hiddenInput);
-            initJsTreeAjax('#container-folder-form-tree', "multi-select-folder", true, 600000);
-        }
-
-        function triggerWebhookMultiSelectFormOpen(dialog) {
-            var $triggerCode =   $(
-                '<div class="form-group row">' +
-                '<label class="col-sm-2 control-label" for="triggerCode" accesskey="">{% trans "Trigger Code" %}</label>' +
-                '<div class="col-sm-10">' +
-                '<input class="form-control" name="triggerCode" type="text" id="triggerCode" value="">' +
-                '<span class="help-block">{% trans "Enter the code associated with the web hook you wish to trigger. Please note that for this action to work, the webhook trigger code has to be added to Interactive Actions in scheduled content for this Player."%}</span>' +
-                '</div>' +
-                '</div>'
-            );
-
-            $(dialog).find('.modal-body').append($triggerCode);
-
-            $('#triggerCode').on('change', function() {
-
-                dialog.data().commitData = {
-                    triggerCode: $('#triggerCode').val()
-                };
-            });
-        }
-
-        function sendCommandMultiSelectFormOpen(dialog) {
-          // Inject a list of commands into the form, in a drop down.
-          const $commandSelect = $(
-            '<div class="form-group form-horizontal row mt-4">' +
-            '<label class="col-sm-2 control-label" for="commandId" accesskey="">{% trans "Command" %}</label>' +
-            '<div class="col-sm-10">' +
-            '<select name="commandId" class="form-control" data-search-url="{{ url_for("command.search") }}" data-search-term="command" data-id-property="commandId" data-text-property="command">' +
-            '<span class="help-block">{% trans "Pick a command to send to the Player. If the CMS has XMR enabled this will be sent immediately, otherwise it will show an error."%}</span>' +
-            '</div>' +
-            '</div>'
-          );
-
-          // Add the list to the body.
-          $(dialog).find('.modal-body').append($commandSelect);
-
-          makePagedSelect(dialog.find('select[name="commandId"]'), dialog);
-
-          dialog.find('select[name="commandId"]').on('select2:select', function() {
-            dialog.data().commitData = {commandId: $(this).select2('data')[0].id};
-          });
-        }
-    </script>
-
     {# File upload templates and scripts #}
     {% include "include-file-upload.twig" %}
 {% endblock %}
