+++ views/developer-template-edit-page.twig
@@ -1025,6 +1025,6 @@
         });
     </script>
 
-    {# Add Monaco editor bundle #}
-    <script type="text/javascript" src="{{ theme.rootUri() }}dist/monacoEditor.bundle.min.js?v={{ version }}&rev={{revision}}" nonce="{{ cspNonce }}" defer></script>
+    {# Add code editor bundle #}
+    <script type="text/javascript" src="{{ theme.rootUri() }}dist/codeEditor.bundle.min.js?v={{ version }}&rev={{revision}}" nonce="{{ cspNonce }}" defer></script>
 {% endblock %}
