+++ views/developer-template-page.twig
@@ -91,133 +91,28 @@
 {% endblock %}
 
 {% block javaScript %}
-    <script type="text/javascript" nonce="{{ cspNonce }}">
-      var table = $('#templates').DataTable({
-        language: dataTablesLanguage,
-        dom: dataTablesTemplate,
-        serverSide: true,
-        stateSave: true,
-        stateDuration: 0,
-        responsive: true,
-        stateLoadCallback: dataTableStateLoadCallback,
-        stateSaveCallback: dataTableStateSaveCallback,
-        filter: false,
-        searchDelay: 3000,
-        order: [[ 0, 'asc']],
-        ajax: {
-          url: '{{ url_for("developer.templates.search") }}',
-          data: function (d) {
-            $.extend(d, $('#templates').closest(".XiboGrid").find(".FilterDiv form").serializeObject());
-          }
-        },
-        columns: [
-          { data: 'id' , responsivePriority: 2},
-          { data: 'templateId' , responsivePriority: 2},
-          { data: 'dataType' },
-          { data: 'title', orderable: false },
-          { data: 'type', orderable: false },
-          {
-            data: 'groupsWithPermissions',
-            responsivePriority: 3,
-            render: dataTableCreatePermissions
-          },
-          {
-            orderable: false,
-            responsivePriority: 1,
-            data: dataTableButtonsColumn
-          }
-        ]
-      });
+    {# Initialise JS variables and translations #}
+    <script type="text/javascript" nonce="{{ cspNonce }}" defer>
+        {# JS variables #}
+        var developerTemplatesSearchURL = "{{ url_for('developer.templates.search') }}";
+        var developerTemplatesImportURL = "{{ url_for('developer.templates.import') }}";
+        var moduleTemplateSearchURL = "{{ url_for('module.template.search', {dataType : ':dataType'}) }}";
 
-      table.on('draw', dataTableDraw);
-      table.on('processing.dt', dataTableProcessing);
-      dataTableAddButtons(table, $('#templates_wrapper').find('.dataTables_buttons'));
-
-      $("#refreshGrid").click(function () {
-          table.ajax.reload();
-      });
-
-      $('#module-template-xml-import').click(function(e) {
-        e.preventDefault();
-
-        openUploadForm({
-          url: "{{ url_for("developer.templates.import") }}",
-          templateId : 'template-module-xml-upload',
-          uploadTemplateId: 'template-module-xml-upload-files',
-          title: "{% trans "Import XML" %}",
-          initialisedBy: "module-templates-upload",
-          buttons: {
-            main: {
-              label: "{% trans "Done" %}",
-              className: "btn-primary btn-bb-main",
-              callback: function () {
-                table.ajax.reload();
-                XiboDialogClose();
-              }
-            }
-          },
-          templateOptions: {
-            multi: false,
-            trans: {
+        {# Custom translations #}
+        var developerTemplatePageTrans = {
+            importXML: "{% trans "Import XML" %}",
+            done: "{% trans "Done" %}",
+            templateOptions: {
               addFiles: "{% trans "Add files" %}",
               startUpload: "{% trans "Start upload" %}",
               cancelUpload: "{% trans "Cancel upload" %}",
             },
-            upload: {
-              validExt: "xml"
-            },
-          }
-        });
-      });
-
-      function moduleTemplateAddFormOpen(dialog) {
-        const $form = $(dialog).find('#form-module-template');
-        $('#dataType', $form)
-          .on('select2:select', function(e) {
-            let dataType = $(this).select2('data')[0].id;
-            let $templateSelect = $(dialog).find('#copyTemplateId');
-            let searchUrl =
-              "{{ url_for('module.template.search', {dataType : ':dataType'}) }}"
-                .replace(':dataType', dataType);
-
-            $templateSelect.data('searchUrl', searchUrl);
-            makeTemplateSelect($templateSelect)
-            $templateSelect.parent().parent().removeClass('d-none')
-            $templateSelect.val(null).trigger('change');
-          })
-      }
-
-      function makeTemplateSelect($element) {
-        // clear existing options.
-        $element.empty().trigger('change');
-        // append empty option
-        $element.append(new Option('', '', false, false));
-
-        // get static templates for the selected dataType
-        $.ajax({
-          method: 'GET',
-          url: $element.data('searchUrl'),
-          data: $element.data('filterOptions'),
-          dataType: 'json',
-          success: function(response) {
-            $.each(response.data, function(key, el) {
-              $element.append(new Option(el.title, el.templateId, false, false));
-            });
-
-            $element.select2({
-              allowClear: true,
-              placeholder: {
-                id: null,
-                value: ''
-              },
-            });
-          },
-          error: function(xhr) {
-            SystemMessage(xhr.message || '{{"Unknown error"|trans}}', false);
-          }
-        });
-      }
+            unknownError: "{% trans "Unknown error" %}",
+        };
     </script>
+
+    {# Add page source code bundle ( JS and CSS ) #}
+    <script src="{{ theme.rootUri() }}dist/pages/developer-template-page.bundle.min.js?v={{ version }}&rev={{revision}}" nonce="{{ cspNonce }}"></script>
 {% endblock %}
 
 {% block javaScriptTemplates %}
