+++ views/include-file-upload.twig
@@ -13,7 +13,7 @@
 <script src="{{ theme.rootUri() }}dist/vendor/jquery-file-upload/js/jquery.fileupload-audio.js?v={{ version }}&rev={{revision}}" nonce="{{ cspNonce }}"></script>
 
 {# Our implementation of blueimp #}
-<script src="{{ theme.rootUri() }}dist/core/file-upload.js?v={{ version }}&rev={{revision}}" nonce="{{ cspNonce }}"></script>
+<script src="{{ theme.rootUri() }}dist/fileUpload.bundle.min.js?v={{ version }}&rev={{revision}}" nonce="{{ cspNonce }}"></script>
 
 {# Max image resize size from settings #}
 <script type="text/javascript" nonce="{{ cspNonce }}">
