--- views/module-form-delete.twig
@@ -1,57 +1,0 @@
-{#
-/**
- * Copyright (C) 2022 Xibo Signage Ltd
- *
- * Xibo - Digital Signage - http://www.xibo.org.uk
- *
- * This file is part of Xibo.
- *
- * Xibo is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * any later version.
- *
- * Xibo is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with Xibo.  If not, see <http://www.gnu.org/licenses/>.
- */
-#}
-
-{% extends "form-base.twig" %}
-{% import "forms.twig" as forms %}
-
-{% block formTitle %}
-    {% trans "Delete Widget" %}
-{% endblock %}
-
-{% block formButtons %}
-    {% trans "Cancel" %}, XiboDialogClose()
-    {% trans "Yes" %}, $("#widgetDeleteForm").submit()
-{% endblock %}
-
-{% block formHtml %}
-    <div class="row">
-        <div class="col-md-12">
-            <form id="widgetDeleteForm" class="XiboForm form-horizontal" method="delete" action="{{ url_for("module.widget.delete", {id: widgetId}) }}">
-                {% set message %}{% trans "Are you sure you want to remove this widget?" %}{% endset %}
-                {{ forms.message(message) }}
-
-                {% set message %}{% trans "This action cannot be undone." %}{% endset %}
-                {{ forms.message(message) }}
-
-                {% if module.getModule().regionSpecific == 0 and module.widget.mediaIds|length > 0 %}
-                    {% set message %}{% trans "This widget is linked to Media in the Library which is also in use elsewhere. If the Library item should be deleted then please do so from the Library page." %}{% endset %}
-                    {{ forms.message(message, "alert alert-info") }}
-                {% endif %}
-
-                {% if error %}
-                    <p class="alert alert-danger">{{ "There is an error the module used by your widget, we advise that you delete it."|trans }}</p>
-                {% endif %}
-            </form>
-        </div>
-    </div>
-{% endblock %}
