--- views/region-form-delete.twig
@@ -1,45 +1,0 @@
-{#
-/**
- * Copyright (C) 2020 Xibo Signage Ltd
- *
- * Xibo - Digital Signage - http://www.xibo.org.uk
- *
- * This file is part of Xibo.
- *
- * Xibo is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * any later version.
- *
- * Xibo is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with Xibo.  If not, see <http://www.gnu.org/licenses/>.
- */
-#}
-
-{% extends "form-base.twig" %}
-{% import "forms.twig" as forms %}
-
-{% block formTitle %}
-    {% trans "Delete this region?" %}
-{% endblock %}
-
-{% block formButtons %}
-    {% trans "Cancel" %}, XiboDialogClose()
-    {% trans "Yes" %}, DeleteRegion($("#regionFormDelete")| {{ region.regionId }})
-{% endblock %}
-
-{% block formHtml %}
-    <div class="row">
-        <div class="col-md-12">
-            <form id="regionFormDelete" class="XiboForm form-horizontal" method="delete" action="{{ url_for("region.delete", {id: region.regionId}) }}">
-                {% set message %}{% trans "Are you sure you want to remove this region? All media files will be unassigned and any context saved to the region itself (such as Text, Tickers) will be lost permanently." %}{% endset %}
-                {{ forms.message(message) }}
-            </form>
-        </div>
-    </div>
-{% endblock %}
